refactor(agent): extract session/error-handling mixins from run_agent.py (shard s2)
