+++ plugins/agent/__init__.py
@@ -1,0 +1,1 @@
+# Agent mixins package (run_agent.py extraction)
