+++ plugins/agent/mixins/__init__.py
@@ -1,0 +1,1 @@
+# Mixin modules extracted verbatim from run_agent.py (wave 1)
