+++ run_agent.py
@@ -61,7 +61,6 @@ from typing import List, Dict, Any, Optional, Callable
 # ModuleNotFoundError on broken/partial installs where `fire` isn't present.
 from datetime import datetime
 from pathlib import Path
-from types import SimpleNamespace
 
 from hermes_constants import get_hermes_home
 
@@ -157,7 +156,6 @@ from agent.model_metadata import (
     estimate_request_tokens_rough,  # noqa: F401  # re-exported for tests that mock.patch("run_agent.estimate_request_tokens_rough")
     is_local_endpoint,
 )
-from agent.usage_pricing import normalize_usage
 # Re-exported for tests that monkeypatch these symbols on run_agent.
 from agent.context_compressor import (  # noqa: F401
     COMPRESSED_SUMMARY_METADATA_KEY,
@@ -205,7 +203,6 @@ from agent.tool_result_classification import (
 )
 from agent.trajectory import (
     convert_scratchpad_to_think,
-    save_trajectory as _save_trajectory_to_file,
 )
 from agent.tool_dispatch_helpers import (
     _should_parallelize_tool_batch,  # noqa: F401  # re-exported for tests that `from run_agent import _should_parallelize_tool_batch`
@@ -221,6 +218,11 @@ from agent.tool_dispatch_helpers import (
     _trajectory_normalize_msg,  # noqa: F401  # re-exported for tests that `from run_agent import _trajectory_normalize_msg`
 )
 from utils import atomic_json_write, base_url_host_matches, base_url_hostname, env_float, is_truthy_value, model_forces_max_completion_tokens
+from plugins.agent.mixins.api_error_mixin import ApiErrorMixin
+from plugins.agent.mixins.api_hooks_mixin import ApiHooksMixin
+from plugins.agent.mixins.background_review_mixin import BackgroundReviewMixin
+from plugins.agent.mixins.reasoning_heuristics_mixin import ReasoningHeuristicsMixin
+from plugins.agent.mixins.trajectory_mixin import TrajectoryMixin
 
 
 # Internal flags that mark a message as ephemeral empty-response/prefill
@@ -409,7 +411,7 @@ class _StreamErrorEvent(Exception):
         }
 
 
-class AIAgent:
+class AIAgent(ReasoningHeuristicsMixin, ApiHooksMixin, BackgroundReviewMixin, ApiErrorMixin, TrajectoryMixin):
     """
     AI Agent with tool calling capabilities.
 
@@ -1642,125 +1644,6 @@ class AIAgent:
                 return value
         return None
 
-    def _has_content_after_think_block(self, content: str) -> bool:
-        """
-        Check if content has actual text after any reasoning/thinking blocks.
-
-        This detects cases where the model only outputs reasoning but no actual
-        response, which indicates an incomplete generation that should be retried.
-        Must stay in sync with _strip_think_blocks() tag variants.
-
-        Args:
-            content: The assistant message content to check
-
-        Returns:
-            True if there's meaningful content after think blocks, False otherwise
-        """
-        if not content:
-            return False
-
-        # Remove all reasoning tag variants (must match _strip_think_blocks)
-        cleaned = self._strip_think_blocks(content)
-
-        # Check if there's any non-whitespace content remaining
-        return bool(cleaned.strip())
-
-    def _strip_think_blocks(self, content: str) -> str:
-        """Forwarder — see ``agent.agent_runtime_helpers.strip_think_blocks``."""
-        from agent.agent_runtime_helpers import strip_think_blocks
-        return strip_think_blocks(self, content)
-
-    @staticmethod
-    def _has_natural_response_ending(content: str) -> bool:
-        """Heuristic: does visible assistant text look intentionally finished?"""
-        if not content:
-            return False
-        stripped = content.rstrip()
-        if not stripped:
-            return False
-        if stripped.endswith("```"):
-            return True
-        if stripped.endswith('^'):
-            return True
-        last = stripped[-1]
-        if last in '.!?:)"\']}。！？：）】」』》^':
-            return True
-        # Emoji ranges (Misc Symbols, Dingbats, Emoticons, Supplemental, etc.)
-        if ord(last) >= 0x1F300:
-            return True
-        return False
-
-    def _is_ollama_glm_backend(self) -> bool:
-        """Detect Ollama-hosted GLM models affected by stop misreports.
-
-        Ollama can misreport truncated output as finish_reason='stop'.
-        Detection relies on explicit Ollama signatures:
-        - Port 11434 (Ollama default)
-        - "ollama" in the base URL (e.g. ollama.local, /ollama/ path)
-        - provider explicitly set to "ollama"
-
-        Crucially it does NOT match arbitrary local/private endpoints
-        (LiteLLM/sglang/vLLM/LM Studio proxies, Tailscale boxes), which
-        report finish_reason correctly and were the source of #13971's
-        false-positive truncation continuations.
-        """
-        model_lower = (self.model or "").lower()
-        provider_lower = (self.provider or "").lower()
-        if "glm" not in model_lower and provider_lower != "zai":
-            return False
-        if "ollama" in self._base_url_lower or ":11434" in self._base_url_lower:
-            return True
-        return provider_lower == "ollama"
-
-    def _should_treat_stop_as_truncated(
-        self,
-        finish_reason: str,
-        assistant_message,
-        messages: Optional[list] = None,
-    ) -> bool:
-        """Detect conservative stop->length misreports for Ollama-hosted GLM models."""
-        if finish_reason != "stop" or self.api_mode != "chat_completions":
-            return False
-        if not self._is_ollama_glm_backend():
-            return False
-        if not any(
-            isinstance(msg, dict) and msg.get("role") == "tool"
-            for msg in (messages or [])
-        ):
-            return False
-        if assistant_message is None or getattr(assistant_message, "tool_calls", None):
-            return False
-
-        content = getattr(assistant_message, "content", None)
-        if not isinstance(content, str):
-            return False
-
-        visible_text = self._strip_think_blocks(content).strip()
-        if not visible_text:
-            return False
-        if len(visible_text) < 20 or not re.search(r"\s", visible_text):
-            return False
-
-        return not self._has_natural_response_ending(visible_text)
-
-    def _looks_like_codex_intermediate_ack(
-        self,
-        user_message: str,
-        assistant_content: str,
-        messages: List[Dict[str, Any]],
-        require_workspace: bool = True,
-    ) -> bool:
-        """Forwarder — see ``agent.agent_runtime_helpers.looks_like_codex_intermediate_ack``."""
-        from agent.agent_runtime_helpers import looks_like_codex_intermediate_ack
-        return looks_like_codex_intermediate_ack(
-            self, user_message, assistant_content, messages, require_workspace
-        )
-
-    def _extract_reasoning(self, assistant_message) -> Optional[str]:
-        """Forwarder — see ``agent.agent_runtime_helpers.extract_reasoning``."""
-        from agent.agent_runtime_helpers import extract_reasoning
-        return extract_reasoning(self, assistant_message)
-
     def _cleanup_task_resources(self, task_id: str) -> None:
         """Forwarder — see ``agent.chat_completion_helpers.cleanup_task_resources``."""
         from agent.chat_completion_helpers import cleanup_task_resources
@@ -1774,67 +1657,6 @@ class AIAgent:
         _SKILL_REVIEW_PROMPT,
         _COMBINED_REVIEW_PROMPT,
     )
-
-    @staticmethod
-    def _summarize_background_review_actions(
-        review_messages: List[Dict],
-        prior_snapshot: List[Dict],
-        notification_mode: str = "on",
-    ) -> List[str]:
-        """Forwarder — see ``agent.background_review.summarize_background_review_actions``."""
-        from agent.background_review import summarize_background_review_actions
-        return summarize_background_review_actions(
-            review_messages,
-            prior_snapshot,
-            notification_mode=notification_mode,
-        )
-
-    def _spawn_background_review(
-        self,
-        messages_snapshot: List[Dict],
-        review_memory: bool = False,
-        review_skills: bool = False,
-    ) -> None:
-        """Spawn the background memory/skill review thread.
-
-        Thin wrapper — the heavy lifting lives in
-        ``agent.background_review.spawn_background_review_thread`` which
-        returns the thread target.  ``threading.Thread`` is constructed
-        here so existing tests that patch ``run_agent.threading.Thread``
-        keep working.
-        """
-        from agent.background_review import spawn_background_review_thread
-        from tools.thread_context import propagate_context_to_thread
-        target, _prompt = spawn_background_review_thread(
-            self,
-            messages_snapshot,
-            review_memory=review_memory,
-            review_skills=review_skills,
-        )
-        # Carry the active profile into the review thread so MEMORY.md / skill
-        # review writes land in the right profile (#54937).
-        t = threading.Thread(
-            target=propagate_context_to_thread(target), daemon=True, name="bg-review"
-        )
-        t.start()
-
-    def _build_memory_write_metadata(
-        self,
-        *,
-        write_origin: Optional[str] = None,
-        execution_context: Optional[str] = None,
-        task_id: Optional[str] = None,
-        tool_call_id: Optional[str] = None,
-    ) -> Dict[str, Any]:
-        """Forwarder — see ``agent.background_review.build_memory_write_metadata``."""
-        from agent.background_review import build_memory_write_metadata
-        return build_memory_write_metadata(
-            self,
-            write_origin=write_origin,
-            execution_context=execution_context,
-            task_id=task_id,
-            tool_call_id=tool_call_id,
-        )
 
     def _apply_persist_user_message_override(self, messages: List[Dict]) -> None:
         """Rewrite the current-turn user message before persistence/return.
@@ -2306,582 +2128,6 @@ class AIAgent:
         
         # Return everything up to (not including) the last assistant message
         return messages[:last_assistant_idx]
-
-    def _format_tools_for_system_message(self) -> str:
-        """Forwarder — see ``agent.system_prompt.format_tools_for_system_message``."""
-        from agent.system_prompt import format_tools_for_system_message
-        return format_tools_for_system_message(self)
-
-    def _convert_to_trajectory_format(self, messages: List[Dict[str, Any]], user_query: str, completed: bool) -> List[Dict[str, Any]]:
-        """Forwarder — see ``agent.agent_runtime_helpers.convert_to_trajectory_format``."""
-        from agent.agent_runtime_helpers import convert_to_trajectory_format
-        return convert_to_trajectory_format(self, messages, user_query, completed)
-
-    def _save_trajectory(self, messages: List[Dict[str, Any]], user_query: str, completed: bool):
-        """
-        Save conversation trajectory to JSONL file.
-        
-        Args:
-            messages (List[Dict]): Complete message history
-            user_query (str): Original user query
-            completed (bool): Whether the conversation completed successfully
-        """
-        if not self.save_trajectories:
-            return
-        
-        trajectory = self._convert_to_trajectory_format(messages, user_query, completed)
-        _save_trajectory_to_file(trajectory, self.model, completed)
-
-    @staticmethod
-    def _is_entitlement_failure(
-        error_context: Optional[Dict[str, Any]],
-        status_code: Optional[int],
-    ) -> bool:
-        """Detect subscription/entitlement 403s that masquerade as auth failures.
-
-        Returned True only when the body text matches a known entitlement
-        shape AND the status is 401/403.  Refreshing an OAuth token cannot
-        fix an unsubscribed account, so callers should surface the error
-        instead of looping the credential pool.
-
-        Current matches:
-          * xAI OAuth: "do not have an active Grok subscription" /
-            "out of available resources" / "does not have permission" + "grok"
-
-        Disambiguator for xAI (#29344): the same ``code`` text ("The caller
-        does not have permission to execute the specified operation") is
-        returned for BOTH an unsubscribed account AND a stale OAuth access
-        token.  xAI ships an explicit signal in the ``error`` field that
-        tells the two apart: a ``[WKE=unauthenticated:...]`` suffix (and/or
-        the ``OAuth2 access token could not be validated`` phrasing) means
-        the credentials failed validation — that's recoverable by refreshing
-        the token, NOT by surfacing an entitlement message.  When either
-        signal is present we return False eagerly so the credential-pool
-        refresh path runs, letting long-running TUI sessions recover from
-        stale tokens without an exit/reopen cycle.
-
-        Extend here for new providers as we discover them (Anthropic's
-        Claude Max OAuth entitlement errors look distinct enough today that
-        the existing 1M-context-beta branch handles them; revisit if other
-        subscription tiers start producing the same loop signature).
-        """
-        if status_code not in {401, 403, None}:
-            return False
-        if not isinstance(error_context, dict):
-            return False
-        # Build a single lowercase haystack covering every field shape the
-        # body might land in.  ``_extract_api_error_context`` normalises to
-        # ``message``/``reason``, but callers (and the test suite) may also
-        # hand us the raw body with ``code``/``error`` keys; cover both so
-        # the WKE disambiguator below fires regardless of entry point.
-        message = str(error_context.get("message") or "").lower()
-        reason = str(error_context.get("reason") or "").lower()
-        code = str(error_context.get("code") or "").lower()
-        err = str(error_context.get("error") or "").lower()
-        haystack = f"{message} {reason} {code} {err}"
-        if not haystack.strip():
-            return False
-        # xAI's authoritative disambiguator for "stale token" vs
-        # "unsubscribed account".  Both conditions share the same
-        # permission-denied ``code`` text; only one carries this suffix.
-        # Bail out before the entitlement keyword checks so a stale OAuth
-        # token routes through the credential-refresh path instead of the
-        # surface-error-as-entitlement path.  See #29344 for the long-
-        # running TUI failure mode this closes.
-        if "[wke=unauthenticated:" in haystack:
-            return False
-        if "oauth2 access token could not be validated" in haystack:
-            return False
-        if "do not have an active grok subscription" in haystack:
-            return True
-        if "out of available resources" in haystack and "grok" in haystack:
-            return True
-        if "does not have permission" in haystack and "grok" in haystack:
-            return True
-        return False
-
-    @staticmethod
-    def _decorate_xai_entitlement_error(detail: str) -> str:
-        """Append a neutral hint when xAI's OAuth surface returns the
-        permission-denied 403.
-
-        xAI's ``/v1/responses`` endpoint replies to several distinct failure
-        modes with the SAME body::
-
-            {"code": "The caller does not have permission to execute the
-             specified operation", "error": "You have either run out of
-             available resources or do not have an active Grok subscription.
-             Manage subscriptions at https://grok.com/?_s=usage or subscribe
-             at https://grok.com/supergrok"}
-
-        That body covers several real causes we cannot distinguish without
-        more info from xAI.  The most common (and least obvious) one is
-        that **X Premium+ does NOT include API access** — only standalone
-        SuperGrok subscribers can use Hermes against xai-oauth.  Lots of
-        users see Grok in their X app, assume it works here too, and hit
-        this 403 with no idea why.  Lead the hint with that.
-
-        Other possible causes:
-          * No Grok subscription at all
-          * SuperGrok tier doesn't include the requested model (e.g.
-            grok-4.3 may need a higher tier)
-          * Monthly quota exhausted (the ``?_s=usage`` URL hints at this)
-
-        Surface the raw xAI text verbatim and point at
-        https://grok.com/?_s=usage where the user can see WHICH applies.
-
-        Matched once per detail string — won't double-decorate if the
-        upstream already concatenated the same text.
-        """
-        if not detail:
-            return detail
-        lower = detail.lower()
-        is_entitlement = (
-            "do not have an active grok subscription" in lower
-            or ("out of available resources" in lower and "grok" in lower)
-            or ("does not have permission" in lower and "grok" in lower)
-        )
-        if not is_entitlement:
-            return detail
-        hint = (
-            " — xAI rejected this OAuth account. NOTE: X Premium+ does NOT "
-            "include xAI API access — only standalone SuperGrok subscribers "
-            "can use this provider. Other possible causes: no Grok "
-            "subscription, your tier doesn't include this model, or your "
-            "quota is exhausted. Check https://grok.com/?_s=usage to see "
-            "which, or run `/model` to switch providers."
-        )
-        # Idempotency: detect prior decoration by a substring unique to the
-        # hint (not present in xAI's own body text).
-        if "X Premium+ does NOT include" in detail:
-            return detail
-        return f"{detail}{hint}"
-
-    @staticmethod
-    def _coerce_api_error_detail(value: Any) -> str:
-        """Return a display-safe string for structured provider error fields."""
-        if isinstance(value, str):
-            return value
-        if isinstance(value, dict):
-            for key in ("message", "detail", "error", "code", "type"):
-                nested = value.get(key)
-                if isinstance(nested, str) and nested.strip():
-                    return nested
-            for key in ("message", "detail", "error", "code", "type"):
-                if key in value:
-                    nested_detail = AIAgent._coerce_api_error_detail(value[key])
-                    if nested_detail:
-                        return nested_detail
-            try:
-                return json.dumps(value, ensure_ascii=False, sort_keys=True)
-            except TypeError:
-                return str(value)
-        if isinstance(value, (list, tuple)):
-            parts = [
-                AIAgent._coerce_api_error_detail(item)
-                for item in value
-            ]
-            return "; ".join(part for part in parts if part)
-        if value is None:
-            return ""
-        return str(value)
-
-    @staticmethod
-    def _summarize_api_error(error: Exception) -> str:
-        """Extract a human-readable one-liner from an API error.
-
-        Handles Cloudflare HTML error pages (502, 503, etc.) by pulling the
-        <title> tag instead of dumping raw HTML.  Falls back to a truncated
-        str(error) for everything else.
-        """
-        raw = str(error)
-
-        if (
-            isinstance(error, ValueError)
-            and "expected ident at line" in raw.lower()
-        ):
-            return f"Malformed provider streaming response: {raw[:300]}"
-
-        # Cloudflare / proxy HTML pages: grab the <title> for a clean summary
-        if "<!DOCTYPE" in raw or "<html" in raw:
-            m = re.search(r"<title[^>]*>([^<]+)</title>", raw, re.IGNORECASE)
-            title = m.group(1).strip() if m else "HTML error page (title not found)"
-            # Also grab Cloudflare Ray ID if present
-            ray = re.search(r"Cloudflare Ray ID:\s*<strong[^>]*>([^<]+)</strong>", raw)
-            ray_id = ray.group(1).strip() if ray else None
-            status_code = getattr(error, "status_code", None)
-            parts = []
-            if status_code:
-                parts.append(f"HTTP {status_code}")
-            parts.append(title)
-            if ray_id:
-                parts.append(f"Ray {ray_id}")
-            return " — ".join(parts)
-
-        # GeminiAPIError (agent/gemini_native_adapter.py) already composes a
-        # clean one-liner and may have appended actionable guidance (free-tier
-        # 429, legacy Standard-key 401). Prefer its message over re-extracting
-        # the raw response body below, which would strip that guidance.
-        if type(error).__name__ == "GeminiAPIError":
-            return redact_sensitive_text(raw[:1000])
-
-        # JSON body errors from OpenAI/Anthropic SDKs
-        body = getattr(error, "body", None)
-        if isinstance(body, dict):
-            msg = body.get("error", {}).get("message") if isinstance(body.get("error"), dict) else body.get("message")
-            if msg:
-                status_code = getattr(error, "status_code", None)
-                prefix = f"HTTP {status_code}: " if status_code else ""
-                msg = AIAgent._coerce_api_error_detail(msg)
-                return AIAgent._decorate_xai_entitlement_error(f"{prefix}{msg[:300]}")
-
-        # SDK may leave body empty while httpx still has the payload (#36109).
-        # Redact before returning: the raw provider/proxy error body is
-        # attacker-influenced and may echo Authorization / x-api-key / request
-        # JSON, which would otherwise leak into final_response + logs (this path
-        # widens exposure vs the old empty-body "HTTP 400" string).
-        response = getattr(error, "response", None)
-        if response is not None:
-            try:
-                snippet = (getattr(response, "text", None) or "").strip()
-            except Exception:
-                snippet = ""
-            if snippet:
-                status_code = getattr(error, "status_code", None)
-                prefix = f"HTTP {status_code}: " if status_code else ""
-                try:
-                    payload = json.loads(snippet)
-                except (json.JSONDecodeError, TypeError):
-                    payload = None
-                if isinstance(payload, dict):
-                    err = payload.get("error")
-                    if isinstance(err, dict) and err.get("message"):
-                        return redact_sensitive_text(f"{prefix}{str(err['message'])[:300]}")
-                    if payload.get("message"):
-                        return redact_sensitive_text(f"{prefix}{str(payload['message'])[:300]}")
-                return redact_sensitive_text(f"{prefix}{snippet[:300]}")
-
-        # Fallback: truncate the raw string but give more room than 200 chars
-        status_code = getattr(error, "status_code", None)
-        prefix = f"HTTP {status_code}: " if status_code else ""
-        return AIAgent._decorate_xai_entitlement_error(f"{prefix}{raw[:500]}")
-
-    def _mask_api_key_for_logs(self, key: Any) -> Optional[str]:
-        # Azure Foundry Entra ID bearer providers are callables — never
-        # invoke them in log paths; identify the auth surface instead.
-        if callable(key) and not isinstance(key, str):
-            return "<entra-id-bearer>"
-        if not key:
-            return None
-        if len(key) <= 12:
-            return "***"
-        return f"{key[:8]}...{key[-4:]}"
-
-    def _clean_error_message(self, error_msg: str) -> str:
-        """
-        Clean up error messages for user display, removing HTML content and truncating.
-        
-        Args:
-            error_msg: Raw error message from API or exception
-            
-        Returns:
-            Clean, user-friendly error message
-        """
-        if not error_msg:
-            return "Unknown error"
-            
-        # Remove HTML content (common with CloudFlare and gateway error pages)
-        if error_msg.strip().startswith('<!DOCTYPE html') or '<html' in error_msg:
-            return "Service temporarily unavailable (HTML error page returned)"
-            
-        # Remove newlines and excessive whitespace
-        cleaned = ' '.join(error_msg.split())
-        
-        # Truncate if too long
-        if len(cleaned) > 150:
-            cleaned = cleaned[:150] + "..."
-            
-        return cleaned
-
-    @staticmethod
-    def _extract_api_error_context(error: Exception) -> Dict[str, Any]:
-        """Forwarder — see ``agent.agent_runtime_helpers.extract_api_error_context``."""
-        from agent.agent_runtime_helpers import extract_api_error_context
-        return extract_api_error_context(error)
-
-    def _usage_summary_for_api_request_hook(self, response: Any) -> Optional[Dict[str, Any]]:
-        """Token buckets for ``post_api_request`` plugins (no raw ``response`` object)."""
-        if response is None:
-            return None
-        raw_usage = getattr(response, "usage", None)
-        if not raw_usage:
-            return None
-        from dataclasses import asdict
-
-        cu = normalize_usage(raw_usage, provider=self.provider, api_mode=self.api_mode)
-        summary = asdict(cu)
-        summary.pop("raw_usage", None)
-        summary["prompt_tokens"] = cu.prompt_tokens
-        summary["total_tokens"] = cu.total_tokens
-        return summary
-
-    @staticmethod
-    def _hook_payload_max_chars() -> int:
-        raw = os.getenv("HERMES_PLUGIN_PAYLOAD_MAX_CHARS", "50000")
-        try:
-            return max(1000, int(raw))
-        except (TypeError, ValueError):
-            return 50000
-
-    @staticmethod
-    def _is_sensitive_hook_key(key: Any) -> bool:
-        if not isinstance(key, str):
-            return False
-        lowered = key.lower().replace("-", "_")
-        exact = {
-            "api_key",
-            "authorization",
-            "proxy_authorization",
-            "cookie",
-            "set_cookie",
-        }
-        return lowered in exact or lowered.endswith("_api_key")
-
-    @classmethod
-    def _hook_jsonable(
-        cls,
-        value: Any,
-        *,
-        depth: int = 0,
-        max_depth: int = 8,
-        max_string: int = 8000,
-        max_sequence: int = 200,
-    ) -> Any:
-        if depth > max_depth:
-            return f"<{type(value).__name__} depth limit>"
-        if value is None or isinstance(value, (bool, int, float)):
-            return value
-        if isinstance(value, str):
-            if len(value) > max_string:
-                return value[:max_string] + f"...[truncated {len(value) - max_string} chars]"
-            return value
-        if isinstance(value, (bytes, bytearray)):
-            return f"<{len(value)} bytes>"
-        if isinstance(value, dict):
-            out: Dict[str, Any] = {}
-            for idx, (key, item) in enumerate(value.items()):
-                if idx >= max_sequence:
-                    out["_truncated_items"] = len(value) - max_sequence
-                    break
-                str_key = str(key)
-                if cls._is_sensitive_hook_key(str_key):
-                    out[str_key] = "<redacted>"
-                else:
-                    out[str_key] = cls._hook_jsonable(
-                        item,
-                        depth=depth + 1,
-                        max_depth=max_depth,
-                        max_string=max_string,
-                        max_sequence=max_sequence,
-                    )
-            return out
-        if isinstance(value, (list, tuple, set)):
-            seq = list(value)
-            out = [
-                cls._hook_jsonable(
-                    item,
-                    depth=depth + 1,
-                    max_depth=max_depth,
-                    max_string=max_string,
-                    max_sequence=max_sequence,
-                )
-                for item in seq[:max_sequence]
-            ]
-            if len(seq) > max_sequence:
-                out.append({"_truncated_items": len(seq) - max_sequence})
-            return out
-        try:
-            if hasattr(value, "model_dump"):
-                try:
-                    dumped = value.model_dump(mode="json")
-                except TypeError:
-                    dumped = value.model_dump()
-                return cls._hook_jsonable(
-                    dumped,
-                    depth=depth + 1,
-                    max_depth=max_depth,
-                    max_string=max_string,
-                    max_sequence=max_sequence,
-                )
-        except Exception:
-            pass
-        try:
-            from dataclasses import asdict, is_dataclass
-            if is_dataclass(value):
-                return cls._hook_jsonable(
-                    asdict(value),
-                    depth=depth + 1,
-                    max_depth=max_depth,
-                    max_string=max_string,
-                    max_sequence=max_sequence,
-                )
-        except Exception:
-            pass
-        if isinstance(value, SimpleNamespace):
-            return cls._hook_jsonable(
-                vars(value),
-                depth=depth + 1,
-                max_depth=max_depth,
-                max_string=max_string,
-                max_sequence=max_sequence,
-            )
-        if hasattr(value, "__dict__"):
-            try:
-                public_attrs = {
-                    k: v
-                    for k, v in vars(value).items()
-                    if not str(k).startswith("_")
-                }
-                return cls._hook_jsonable(
-                    public_attrs,
-                    depth=depth + 1,
-                    max_depth=max_depth,
-                    max_string=max_string,
-                    max_sequence=max_sequence,
-                )
-            except Exception:
-                pass
-        return str(value)[:max_string]
-
-    @classmethod
-    def _sanitize_hook_payload(cls, value: Any) -> Any:
-        payload = cls._hook_jsonable(value)
-        limit = cls._hook_payload_max_chars()
-        try:
-            encoded = json.dumps(payload, ensure_ascii=False, default=str)
-        except Exception:
-            return str(payload)[:limit]
-        if len(encoded) <= limit:
-            return payload
-        payload = cls._hook_jsonable(value, max_string=1000, max_sequence=50)
-        try:
-            encoded = json.dumps(payload, ensure_ascii=False, default=str)
-        except Exception:
-            return str(payload)[:limit]
-        if len(encoded) <= limit:
-            return payload
-        return {
-            "_truncated": True,
-            "original_type": type(value).__name__,
-            "preview": encoded[:limit],
-        }
-
-    def _api_request_payload_for_hook(self, api_kwargs: Optional[Dict[str, Any]]) -> Dict[str, Any]:
-        body = {
-            key: value
-            for key, value in (api_kwargs or {}).items()
-            if key not in {"timeout", "http_client"}
-        }
-        return self._sanitize_hook_payload(
-            {
-                "method": "POST",
-                "body": body,
-            }
-        )
-
-    def _api_response_payload_for_hook(
-        self,
-        response: Any,
-        assistant_message: Any,
-        *,
-        finish_reason: Optional[str],
-    ) -> Dict[str, Any]:
-        # ``tool_calls`` is the raw list of provider SDK objects (e.g.
-        # OpenAI ``ChatCompletionMessageToolCall``).  We deliberately hand
-        # the raw objects to ``_sanitize_hook_payload`` and rely on
-        # ``_hook_jsonable`` to normalise them via ``model_dump`` /
-        # ``__dict__`` / dataclass introspection — a future refactor of
-        # the sanitiser MUST preserve that capability or hook subscribers
-        # will receive opaque ``str(obj)`` blobs here.
-        tool_calls = getattr(assistant_message, "tool_calls", None) or []
-        return self._sanitize_hook_payload(
-            {
-                "model": getattr(response, "model", None),
-                "finish_reason": finish_reason,
-                "assistant_message": {
-                    "role": getattr(assistant_message, "role", "assistant"),
-                    "content": getattr(assistant_message, "content", None),
-                    "tool_calls": tool_calls,
-                },
-                "usage": self._usage_summary_for_api_request_hook(response),
-            }
-        )
-
-    def _invoke_api_request_error_hook(
-        self,
-        *,
-        task_id: str,
-        turn_id: str,
-        api_request_id: str,
-        api_call_count: int,
-        api_start_time: float,
-        api_kwargs: Optional[Dict[str, Any]],
-        error_type: str,
-        error_message: str,
-        status_code: Optional[int] = None,
-        retry_count: Optional[int] = None,
-        max_retries: Optional[int] = None,
-        retryable: Optional[bool] = None,
-        reason: Optional[str] = None,
-    ) -> None:
-        # Lazy module import (not from-import) so tests can replace lifecycle
-        # dispatch at this call site. After first call the import is a
-        # ``sys.modules`` dict lookup, so retries don't repay any real cost.
-        try:
-            from hermes_cli import lifecycle as _lifecycle
-
-            if not _lifecycle.has_hook("api_request_error"):
-                return
-            ended_at = time.time()
-            _lifecycle.invoke_hook(
-                "api_request_error",
-                task_id=task_id,
-                turn_id=turn_id,
-                api_request_id=api_request_id,
-                session_id=self.session_id or "",
-                platform=self.platform or "",
-                model=self.model,
-                provider=self.provider,
-                base_url=self.base_url,
-                api_mode=self.api_mode,
-                api_call_count=api_call_count,
-                api_duration=ended_at - api_start_time,
-                started_at=api_start_time,
-                ended_at=ended_at,
-                status_code=status_code,
-                retry_count=retry_count,
-                max_retries=max_retries,
-                retryable=retryable,
-                reason=reason,
-                error={
-                    "type": error_type,
-                    "message": error_message,
-                },
-                request=self._api_request_payload_for_hook(api_kwargs),
-            )
-        except Exception:
-            pass
-
-    def _dump_api_request_debug(
-        self,
-        api_kwargs: Dict[str, Any],
-        *,
-        reason: str,
-        error: Optional[Exception] = None,
-    ) -> Optional[Path]:
-        """Forwarder — see ``agent.agent_runtime_helpers.dump_api_request_debug``."""
-        from agent.agent_runtime_helpers import dump_api_request_debug
-        return dump_api_request_debug(self, api_kwargs, reason=reason, error=error)
 
     @staticmethod
     def _clean_session_content(content: str) -> str:
